Add files via upload
This game is called hanged man and it is a good option to learn programming. You can learn concepts like lists, loops (for, while), handle of strings. Create levels of difficulty. This code is not my. I only added the levels.  The code was written by David Aroesti that was my teacher in the course of python he challenged us  for added some levels for this game and that is the result.
Thanks.